Enum numcodecs_zfp::ZfpCodecError

pub enum ZfpCodecError {

    UnsupportedDtype(AnyArrayDType),
    FixedAccuracyModeIntegerData,
    ExcessiveDimensionality {
        shape: Vec<usize>,
    },
    InvalidExpertMode {
        mode: ZfpCompressionMode,
    },
    HeaderEncodeFailed,
    ZfpEncodeFailed,
    EncodedDataNotBytes {
        dtype: AnyArrayDType,
    },
    EncodedDataNotOneDimensional {
        shape: Vec<usize>,
    },
    HeaderDecodeFailed,
    MismatchedDecodeIntoDtype {
        decoded: AnyArrayDType,
        provided: AnyArrayDType,
    },
    MismatchedDecodeIntoShape {
        decoded: Vec<usize>,
        provided: Vec<usize>,
    },
    DecodeUnknownDtype(u32),
    ZfpDecodeFailed,

}

Errors that may occur when applying the ZfpCodec.

Variants

UnsupportedDtype(AnyArrayDType)

ZfpCodec does not support the dtype

FixedAccuracyModeIntegerData

ZfpCodec does not support the fixed accuracy mode for integer data

ExcessiveDimensionality

ZfpCodec only supports 1-4 dimensional data

Fields

shape: Vec<usize>

The unexpected shape of the data

InvalidExpertMode

ZfpCodec was configured with an invalid expert mode

Fields

mode: ZfpCompressionMode

The unexpected compression mode

HeaderEncodeFailed

ZfpCodec failed to encode the header

ZfpEncodeFailed

ZfpCodec failed to encode the data

EncodedDataNotBytes

ZfpCodec can only decode one-dimensional byte arrays but received an array of a different dtype

Fields

dtype: AnyArrayDType

The unexpected dtype of the encoded array

EncodedDataNotOneDimensional

ZfpCodec can only decode one-dimensional byte arrays but received an array of a different shape

Fields

shape: Vec<usize>

The unexpected shape of the encoded array

HeaderDecodeFailed

ZfpCodec failed to decode the header

MismatchedDecodeIntoDtype

ZfpCodec cannot decode the decoded dtype into the provided array

Fields

decoded: AnyArrayDType

Dtype of the decoded data

provided: AnyArrayDType

Dtype of the provided array into which the data is to be decoded

MismatchedDecodeIntoShape

ZfpCodec cannot decode the decoded array into the provided array of a different shape

Fields

decoded: Vec<usize>

Shape of the decoded data

provided: Vec<usize>

Shape of the provided array into which the data is to be decoded

DecodeUnknownDtype(u32)

ZfpCodec failed to decode the data with the unknown dtype

ZfpDecodeFailed

ZfpCodec failed to decode the data