numcodecs_zfp
Function decompress_into

pub fn decompress_into(
    encoded: &[u8],
    decoded: AnyArrayViewMut<'_>,
) -> Result<(), ZfpCodecError>

Decompress the encoded data into a decoded array using ZFP.

Errors

Errors with

ZfpCodecError::HeaderDecodeFailed if decoding the header failed

ZfpCodecError::DecodeUnknownDtype if the encoded data uses an unknown dtype

ZfpCodecError::MismatchedDecodeIntoArray if the decoded array is of the wrong dtype or shape

ZfpCodecError::ZfpDecodeFailed if an opaque decoding error occurred